+++ src/Orm/Types/BigFloat.php
@@ -10,23 +10,10 @@
 
 namespace Swift\Orm\Types;
 
-use Swift\Orm\Mapping\Definition\Field;
-use Swift\Orm\Dbal\TableQuery;
 
-/**
- * Class BigFloat
- * @package Swift\Orm\Types
- */
 final class BigFloat implements TypeInterface {
 
     public const BIG_FLOAT = 'big_float';
-
-    /**
-     * @inheritDoc
-     */
-    public function getSqlDeclaration( Field $field, TableQuery $query ): string {
-        return sprintf('varchar(%s)', $field->getLength() ?? 128);
-    }
 
     public function transformToPhpValue( mixed $value ): float {
         return (float) $value;