+++ src/Orm/Types/Integer.php
@@ -11,26 +11,16 @@
 namespace Swift\Orm\Types;
 
 
-use Swift\Orm\Mapping\Definition\Field;
-use Swift\Orm\Dbal\TableQuery;
-
 final class Integer implements TypeInterface {
 
     public const INT = 'int';
 
-    /**
-     * @inheritDoc
-     */
-    public function getSqlDeclaration( Field $field, TableQuery $query ): string {
-        return sprintf( 'int(%s)', $field->getLength() ?? 255 );
+    public function transformToPhpValue( mixed $value ): int|null {
+        return $value === null ? null : (int) $value;
     }
 
-    public function transformToPhpValue( mixed $value ): ?int {
-        return is_null($value) ? null : (int) $value;
-    }
-
-    public function transformToDatabaseValue( mixed $value ): ?int {
-        return is_null($value) ? null : (int) $value;
+    public function transformToDatabaseValue( mixed $value ): int|null {
+        return $value === null ? null : (int) $value;
     }
 
     public function getName(): string {