+++ src/Orm/Types/Time.php
@@ -10,23 +10,10 @@
 
 namespace Swift\Orm\Types;
 
-use Swift\Orm\Mapping\Definition\Field;
-use Swift\Orm\Dbal\TableQuery;
 
-/**
- * Class Time
- * @package Swift\Orm\Types
- */
 final class Time implements TypeInterface {
 
     public const TIME = 'time';
-
-    /**
-     * @inheritDoc
-     */
-    public function getSqlDeclaration( Field $field, TableQuery $query ): string {
-        return 'datetime';
-    }
 
     public function transformToPhpValue( mixed $value ): string {
         return (new \DateTime($value))->format('H:i:s');